Rename qscintilla2-qt5.rb [ci skip]

+++ .circleci/script.sh
@@ -19,23 +19,37 @@ set -e
 
 ulimit -n 1024
 
-# fix error: 'libintl.h' file not found
-# build qgis with grass
-brew reinstall gettext
-brew unlink gettext && brew link --force gettext
-
 echo ${CHANGED_FORMULAE}
 
 for f in ${CHANGED_FORMULAE};do
-#  if [[ $(brew list --versions ${f}) ]]; then
-#    echo "Clearing previously installed/cached formula ${f}..."
-#    brew uninstall --force --ignore-dependencies ${f} || true
-#  fi
+  deps=$(brew deps --include-build ${f})
+
+  # fix error: Unable to import PyQt5.QtCore
+  # build qscintilla2
+  if [ "$(echo ${deps} | grep -c 'osgeo-pyqt')" != "0" ];then
+    brew reinstall ${CIRCLE_PROJECT_USERNAME}/${CIRCLE_PROJECT_REPONAME}/osgeo-pyqt
+    brew unlink osgeo-pyqt && brew link osgeo-pyqt --force
+    system "python2", "-c", '"import PyQt5.QtCore"'
+    system "python3", "-c", '"import PyQt5.QtCore"'
+  fi
+
+  # fix error: 'libintl.h' file not found
+  # build qgis with grass
+  if [ "$(echo ${deps} | grep -c 'osgeo-grass')" != "0" ];then
+    brew reinstall gettext
+    brew unlink gettext && brew link --force gettext
+  fi
+
+  # if [[ $(brew list --versions ${f}) ]]; then
+  #   echo "Clearing previously installed/cached formula ${f}..."
+  #   brew uninstall --force --ignore-dependencies ${f} || true
+  # fi
+
   echo "Installing changed formula ${f}..."
   # Default installation flag set
   FLAGS="--build-bottle"
 
-  brew install ${FLAGS} ${TRAVIS_REPO_SLUG}/${f}&
+  brew install ${FLAGS} ${CIRCLE_PROJECT_USERNAME}/${CIRCLE_PROJECT_REPONAME}/${f}&
   PID=$!
   # add progress to ensure Travis doesn't complain about no output
   while true; do
@@ -52,5 +66,5 @@ for f in ${CHANGED_FORMULAE};do
   # does running postinstall mess up the bottle?
   # (mentioned that it is skipped if installing with --build-bottle)
   # brew postinstall ${f}
-  brew test ${TRAVIS_REPO_SLUG}/${f}
+  brew test ${CIRCLE_PROJECT_USERNAME}/${CIRCLE_PROJECT_REPONAME}/${f}
 done
